+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** Fix unclickable arrow in `TimePicker`
 - **[NEW]** `MainContent` and children components to handle layout vertical composition
 
 # v21.2.0 (20/02/2020)